refactor: PersonList를 분리하여 manipulationPeople과 PersonList를 일급 컬렉션으로 만듬
